Splitting endpoints into routes
