+++ api/routes/index.js
@@ -1,0 +1,9 @@
+const contractorRoutes = require('./contractorRoutes');
+const userRoutes = require('./userRoutes');
+const schedulesRoutes = require('./schedulesRoutes');
+
+module.exports = {
+  contractorRoutes,
+  userRoutes,
+  schedulesRoutes,
+};
